requirements.txt
tensorflow
flask
pillow
numpy
logging
pydicom